requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.1
django-appconf==1.0.5
django-autoslug==1.9.9
django-cities==0.6.1
django-cities-light==3.9.2
django-cleanup==8.0.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-image-cropping==1.7
django-js-reverse==0.9.1
django-languages==0.1
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-redis==5.2.0
easy-thumbnails==2.8.5
idna==3.4
kombu==5.3.1
Pillow==9.5.0
progressbar2==4.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
pycountry==22.3.5
python-utils==3.7.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.65.1
typing_extensions==4.7.1
Unidecode==1.3.6
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
